requirements.txt
google-cloud-storage==2.18.2
Flask==3.1.0
gunicorn==23.0.0
soundfile==0.12.1
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
Cython==3.0.11
packaging==24.2
nemo-toolkit==2.0.0
librosa==0.10.2.post1
matplotlib==3.9.2
einops==0.8.0
hydra-core==1.3.2
pytorch-lightning==2.4.0
lhotse==1.28.0
transformers==4.46.3
sentencepiece==0.2.0
pandas==2.2.3
jiwer==3.0.5
pyannote.core==5.0.0
webdataset==0.2.100
editdistance==0.8.1
pyannote.metrics==3.2.1
datasets==3.1.0
wandb==0.18.7
nemo-text-processing==1.1.0